rag_bootcamp/backend/requirements.txt
llama-index
google-cloud-secret-manager
llama-index-readers-json
llama-index-readers-gcs
llama-index-embeddings-google-genai
llama-index-embeddings-vertex-endpoint
llama-index-llms-openai-like
llama-index-llms-google-genai
llama-index-vector-stores-weaviate
llama-index-vector-stores-vertexaivectorsearch
llama-index-retrievers-bm25
openinference-instrumentation-llama-index
opentelemetry-sdk
opentelemetry-proto>=1.12.0
pydantic>=2.10.4
google-cloud-aiplatform>=1.76
uvicorn
fastapi